refactor: move RunnerConf and EvalConf to worker_util next to their Conf base
